New file: plotting_utils to refactor plotting
